Extract CompactionTaskScaler
